fix: declare a template for converting elements between cgo and c++ for private key, g1 and g2 elements
