+++ go-bindings/privatekey.h
@@ -22,7 +22,7 @@ extern "C" {
 typedef void* CPrivateKey;
 
 CPrivateKey CPrivateKeyFromBytes(const void* data, const bool modOrder, bool* didErr);
-CPrivateKey CPrivateKeyAggregate(const void** sks, const size_t len);
+CPrivateKey CPrivateKeyAggregate(void** sks, const size_t len);
 CG1Element CPrivateKeyGetG1Element(const CPrivateKey sk, bool* didErr);
 CG2Element CPrivateKeyGetG2Element(const CPrivateKey sk, bool* didErr);
 CG2Element CPrivateKeyGetG2Power(const CPrivateKey sk, const CG2Element el);
